Get rid of compiler warnings

+++ OpenBrew/OpenBrew/OBHopAdditionTableViewDelegate.m
@@ -122,7 +122,7 @@
       break;
     default:
       [NSException raise:@"Invalid hop addition metric"
-                  format:@"%d", self.hopAdditionMetricToDisplay];
+                  format:@"%d", (int) self.hopAdditionMetricToDisplay];
   }
 }
